Fix Interfaces and Shotguns
It seems the default paths for some images were changed at some point in the game, so I've redirected them to the new location.
Still need to adjust the dimensions to fit the crafted items for some of these, but they no longer throw log errors or have invisible buttons!

Also, the shotguns were just broken. The damage type they referenced didn't exist, which seemed to make enemies immortal when hit, and the projectile was invisible.
Turns out the frames file had an errant 's', so I removed that and it works! I also changed the damage type to default, since the intended file doesn't exist.
The shotgun's sound was changed to the rifle, since the shotgun firing sound doesn't exist in the mod. This prevents the shotgun from just being silent, and could be fixed if the files are created/found at some point. I looked back in some commits, in case it DID exist and was accidentally deleted, but couldn't find it.