FN Temporary(?) Patch
Since FU now has a research system that relies on the order recipes are in an array, replacing the start unlocks outright is a no-go.
I don't know if the game has a way to force-lock recipes in scripts, but that's what I'm looking in to.

It's not ideal, but it makes the progression possible for now!
Also added a failsafe for the telescope, so it automatically learns the matter book when the item is held instead of relying on the default unlocks.